Fix osx: cmd-v when using visual mode selection
When there is a visual mode selection a ```s-v(cmd-v)``` would be expected to paste over the visually selected text.

What currently happens is the visual selection is  deselected and the text is appended to the end of the selection. What should happen is the selected text should be overwritten by the contents of the clipboard.